+++ Dev/Cpp/Effekseer/Effekseer/Effekseer.Instance.cpp
@@ -110,7 +110,7 @@ void Instance::GenerateChildrenInRequired(float currentTime)
 					Matrix43 rootMatrix;
 					rootMatrix.Indentity();
 
-					newInstance->Initialize(this, m_generatedChildrenCount[i], std::max(0.0f, this->m_LivingTime), rootMatrix);
+					newInstance->Initialize(this, m_generatedChildrenCount[i], (int32_t)std::max(0.0f, this->m_LivingTime), rootMatrix);
 				}
 
 				m_generatedChildrenCount[i]++;
@@ -623,13 +623,13 @@ void Instance::Initialize( Instance* parent, int32_t instanceNumber, int32_t par
 
 	if( m_pEffectNode->SoundType == ParameterSoundType_Use )
 	{
-		soundValues.delay = m_pEffectNode->Sound.Delay.getValue( *instanceGlobal );
+		soundValues.delay = (int32_t)m_pEffectNode->Sound.Delay.getValue( *instanceGlobal );
 	}
 
 	// UV
 	if (m_pEffectNode->RendererCommon.UVType == ParameterRendererCommon::UV_ANIMATION)
 	{
-		uvTimeOffset = m_pEffectNode->RendererCommon.UV.Animation.StartFrame.getValue(*instanceGlobal);
+		uvTimeOffset = (int32_t)m_pEffectNode->RendererCommon.UV.Animation.StartFrame.getValue(*instanceGlobal);
 		uvTimeOffset *= m_pEffectNode->RendererCommon.UV.Animation.FrameLength;
 	}
 	
@@ -1345,7 +1345,7 @@ RectF Instance::GetUV() const
 	}
 	else if( m_pEffectNode->RendererCommon.UVType == ParameterRendererCommon::UV_SCROLL )
 	{
-		auto time = m_LivingTime + uvTimeOffset;
+		auto time = (int32_t)m_LivingTime + uvTimeOffset;
 
 		return RectF(
 			uvAreaOffset.X + uvScrollSpeed.X * time,
@@ -1355,7 +1355,7 @@ RectF Instance::GetUV() const
 	}
 	else if (m_pEffectNode->RendererCommon.UVType == ParameterRendererCommon::UV_FCURVE)
 	{
-		auto time = m_LivingTime + uvTimeOffset;
+		auto time = (int32_t)m_LivingTime + uvTimeOffset;
 
 		return RectF(
 			uvAreaOffset.X + m_pEffectNode->RendererCommon.UV.FCurve.Position->X.GetValue(time),
